requirements.txt
Django>=5.1.7
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
mysqlclient
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib